requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-filter==21.1
pytest==6.0.1
PyJWT==1.7.1